feat: Review UX when adding action - MEED-7016 - Meeds-io/MIPs#140
This PR will review the UX of action creatdrawer

Update PR
