+++ portlets/src/main/resources/locale/portlet/Challenges_en.properties
@@ -299,8 +299,10 @@ rule.form.label.actionPlaceholder=Choose an action from the program
 rule.form.label.actionNoData=No action yet
 rule.form.label.actionNotFound=Action not found
 rule.form.label.recurrence=Recurrence
-rule.form.label.stepOne=Tell participants what to do
-rule.form.label.stepTwo=Add conditions to contribute
+rule.form.label.stepOne=Description
+rule.form.label.stepTwo=Conditions
+rule.form.label.stepThree=Rewards
+rule.form.label.stepFour=Publication
 rule.form.label.type.automatic=Automatically
 rule.form.label.type.declarative=Manually
 rule.form.label.after=From
