+++ portlets/src/main/webapp/vue-app/rules/initComponents.js
@@ -29,8 +29,11 @@ import RuleDateInfoChip from './components/detail/RuleDateInfoChip.vue';
 import RuleFavoriteAction from './components/detail/RuleFavoriteAction.vue';
 
 import RuleFormDrawer from './components/drawers/RuleFormDrawer.vue';
+import RuleContentFormDrawer from './components/drawers/RuleContentFormDrawer.vue';
 import RuleFormAutomaticFlow from './components/drawers/RuleFormAutomaticFlow.vue';
 import RuleDetailDrawer from './components/drawers/RuleDetailDrawer.vue';
+import RuleFormConnectorItem from './components/drawers/RuleFormConnectorItem.vue';
+
 import RuleAchievementsDrawer from './components/drawers/RuleAchievementsDrawer.vue';
 import ProgramOwnersDrawer from './components/drawers/ProgramOwnersDrawer.vue';
 import RuleDrawers from './components/RuleDrawers.vue';
@@ -72,8 +75,11 @@ const components = {
   'engagement-center-rule-extensions': RuleExtensions,
 
   'engagement-center-rule-form-drawer': RuleFormDrawer,
+  'engagement-center-rule-content-form-drawer': RuleContentFormDrawer,
   'engagement-center-rule-form-automatic-flow': RuleFormAutomaticFlow,
   'engagement-center-rule-detail-drawer': RuleDetailDrawer,
+  'engagement-center-rule-form-connector-item': RuleFormConnectorItem,
+
   'engagement-center-rule-achievements-drawer': RuleAchievementsDrawer,
   'engagement-center-program-owners-drawer': ProgramOwnersDrawer,
   'engagement-center-rule-drawers': RuleDrawers,
